fixed dictionary encoding in a version-agnostic way
Most forks of this project change is_record(Term, dict, 8)
to is_record(Term, dict, 9) to reflect the fact that apparently
dict records are of size 9 in versions at least greater than
R13B03 (seems to be the earliest complaint about the issue). I
haven't verified that size 8 works prior to R13B03, but I'm guessing
it did at one point.

Assuming dicts will always be represented as tuples where the first
element is the dict atom, this change would be version agnostic.

Note that the tests are a bit coupled to the bert implementation,
but with that you gain some documentation as to what it's actually
encoding/decoding.